Move GitRepository to a repository module
